Move pr-create-archive-for-me-screen to `glam` folder
Move this component to the glam folder to keep the onboarding module
more organized.
