+++ src/app/onboarding/components/glam/create-archive-for-me-screen/create-archive-for-me-screen.component.ts
@@ -1,7 +1,7 @@
 /* @format */
 import { Component, Output, EventEmitter } from '@angular/core';
 import { AccountService } from '@shared/services/account/account.service';
-import { ArchiveCreateEvent } from '../glam/types/archive-types';
+import { ArchiveCreateEvent } from '../types/archive-types';
 
 @Component({
   selector: 'pr-create-archive-for-me-screen',
